Updates to const and word, starting solvers -- JM
const.c and word.cpp have been updated to look for their respective
files in the new directories (.\data\const and .\data\dict,
respectively).

Source code for the Solver utilities has been added to the repo.
Right now, there are just a few headers and some definitions for
the Backtracker.